changed polygon tree css and allowed float cl

+++ context/models.py
@@ -33,7 +33,7 @@ class e_Context(models.Model):
 	hydroFeature = models.ForeignKey('rivercureportal.e_HydroFeature', on_delete=models.CASCADE, null=True, blank=True )
 
 	geomExternalBoundary = models.MultiPolygonField(null=True, blank=True)  		#aka Domain
-	CLExternalBoundary  = models.BigIntegerField(null=True, blank=True)  			#aka Domain's CL, characteristic lenght 
+	CLExternalBoundary  = models.FloatField(null=True, blank=True)  			#aka Domain's CL, characteristic lenght 
 
 	user = models.ForeignKey(User, on_delete=models.CASCADE, null=True) # Owner do contexto
 
@@ -71,7 +71,7 @@ class e_ContextFrictionCoeff(models.Model):
 class e_ContextRefinement(models.Model):
 	context = models.ForeignKey('e_Context', on_delete=models.CASCADE, related_name='context_refinement')
 	geom = models.PolygonField(null=True, blank=True) 					# aka Refinement
-	CL  = models.BigIntegerField(null=True, blank=True)  			#aka Refinement's CL
+	CL  = models.FloatField(null=True, blank=True)  			#aka Refinement's CL
 
 	def __str__(self):
 		return f"{self.context} context refinement"
@@ -79,7 +79,7 @@ class e_ContextRefinement(models.Model):
 class e_ContextAlignment(models.Model):
 	context = models.ForeignKey('e_Context', on_delete=models.CASCADE, related_name='context_alignment')
 	geom = models.LineStringField(null=True, blank=True)   		# aka Alignment
-	CL = models.BigIntegerField(null=True, blank=True)   					# Alignment's CL
+	CL = models.FloatField(null=True, blank=True)   					# Alignment's CL
 
 	def __str__(self):
 		return f"{self.context} context alignment"
